src/requirements.txt
flask
requests
pillow
tensorflow